requirements.txt
snowflake-connector-python>=3.6.0
requests>=2.31.0
python-dotenv>=1.0.0
cryptography>=41.0.0
pyjwt>=2.8.0
